pyproject.toml
[project]
name = "repl-nix-collegecounselorai"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-auth>=2.37.0",
    "google-auth-oauthlib>=1.2.1",
    "langchain>=0.3.13",
    "langgraph>=0.2.60",
    "psycopg2-binary>=2.9.10",
    "streamlit>=1.41.1",
]
